Struct safe_core::client::response_getter::ResponseGetter

pub struct ResponseGetter {
    // some fields omitted
}

ResponseGetter is a lazy evaluated response getter. It will fetch either from local cache or wait for the MessageQueue to notify it of the incoming response from the network.

Methods

impl ResponseGetter

fn new(data_channel: Option<(Sender<DataReceivedEvent>, Receiver<DataReceivedEvent>)>, message_queue: Arc<Mutex<MessageQueue>>, requested_type: DataRequest) -> ResponseGetter

Create a new instance of ResponseGetter

fn get(&self) -> Result<Data, CoreError>

Get either from local cache or (if not available there) get it when it comes from the network as informed by MessageQueue. This is blocking.

fn get_sender(&self) -> Option<&Sender<DataReceivedEvent>>

Extract associated sender. This will help cancel the blocking wait at will if so desired. All that is needed is to extract the sender before doing a get() and then while blocking on get() fire sender.send(::translated_events::DataReceivedEvent::Terminated) to gracefully exit the receiver.