Struct safe_core::client::Client

pub struct Client {
    // some fields omitted
}

The main self-authentication client instance that will interface all the request from high level API's to the actual routing layer and manage all interactions with it. This is essentially a non-blocking Client with upper layers having an option to either block and wait on the returned ResponseGetter for receiving network response or spawn a new thread. The Client itself is however well equipped for parallel and non-blocking PUTs and GETS.

Methods

impl Client

fn create_unregistered_client() -> Result<Client, CoreError>

This is a getter-only Gateway function to the Maidsafe network. It will create an unregistered random clinet, which can do very limited set of operations - eg., a Network-Get

fn create_account(keyword: String, pin: String, password: String) -> Result<Client, CoreError>

This is one of the two Gateway functions to the Maidsafe network, the other being the log_in. This will help create a fresh account for the user in the SAFE-network.

fn log_in(keyword: String, pin: String, password: String) -> Result<Client, CoreError>

This is one of the two Gateway functions to the Maidsafe network, the other being the create_account. This will help log into an already created account for the user in the SAFE-network.

fn set_user_root_directory_id(&mut self, root_dir_id: XorName) -> Result<(), CoreError>

Create an entry for the Root Directory ID for the user into the session packet, encrypt and store it. It will be retireved when the user logs into his account. Root directory ID is necessary to fetch all of user's data as all further data is encoded as meta-information into the Root Directory or one of its subdirectories.

fn get_user_root_directory_id(&self) -> Option<&XorName>

Get User's Root Directory ID if available in session packet used for current login

fn set_configuration_root_directory_id(&mut self, root_dir_id: XorName) -> Result<(), CoreError>

Create an entry for the Maidsafe configuration specific Root Directory ID into the session packet, encrypt and store it. It will be retireved when the user logs into his account. Root directory ID is necessary to fetch all of configuration data as all further data is encoded as meta-information into the config Root Directory or one of its subdirectories.

fn get_configuration_root_directory_id(&self) -> Option<&XorName>

Get Maidsafe specific configuration's Root Directory ID if available in session packet used for current login

fn hybrid_encrypt(&self, data_to_encrypt: &[u8], nonce_opt: Option<&Nonce>) -> Result<Vec<u8>, CoreError>

Combined Asymmectric and Symmetric encryption. The data is encrypted using random Key and IV with Xsalsa-symmetric encryption. Random IV ensures that same plain text produces different cipher-texts for each fresh symmetric encryption. The Key and IV are then asymmetrically enrypted using Public-MAID and the whole thing is then serialised into a single Vec.

fn hybrid_decrypt(&self, data_to_decrypt: &[u8], nonce_opt: Option<&Nonce>) -> Result<Vec<u8>, CoreError>

Reverse of hybrid_encrypt. Refer hybrid_encrypt.

fn get(&mut self, request_for: DataRequest, opt_dst: Option<Authority>) -> Result<ResponseGetter, CoreError>

Get data from the network. This is non-blocking.

fn put(&self, data: Data, opt_dst: Option<Authority>) -> Result<(), CoreError>

Put data onto the network. This is non-blocking.

fn post(&self, data: Data, opt_dst: Option<Authority>) -> Result<(), CoreError>

Post data onto the network

fn delete(&self, data: Data, opt_dst: Option<Authority>) -> Result<(), CoreError>

Delete data from the network

fn get_public_encryption_key(&self) -> Result<&PublicKey, CoreError>

Returns the public encryption key

fn get_secret_encryption_key(&self) -> Result<&SecretKey, CoreError>

Returns the Secret encryption key

fn get_public_signing_key(&self) -> Result<&PublicKey, CoreError>

Returns the Public Signing key

fn get_secret_signing_key(&self) -> Result<&SecretKey, CoreError>

Returns the Secret Signing key

fn add_data_receive_event_observer(&self, data_name: XorName, sender: Sender<DataReceivedEvent>)

Add observers for Data Recieve Events

fn add_operation_failure_event_observer(&self, sender: Sender<OperationFailureEvent>)

Add observers for Operation Failure Events like PutFailure, PostFailure, DeleteFailure, Terminated

fn add_network_event_observer(&self, sender: Sender<NetworkEvent>)

Add observers for Network Events like Connected, Disconnected, Terminated

fn get_default_client_manager_address(&self) -> Result<&XorName, CoreError>

Get the default address where the PUTs will go to for this client

fn set_default_client_manager_address(&mut self, address: XorName) -> Result<(), CoreError>

Set the default address where the PUTs and DELETEs will go to for this client