Enum safe_core::errors::CoreError

pub enum CoreError {
    StructuredDataHeaderSizeProhibitive,
    UnsuccessfulEncodeDecode,
    AsymmetricDecipherFailure,
    SymmetricDecipherFailure,
    ReceivedUnexpectedData,
    VersionCacheMiss,
    RoutingMessageCacheMiss,
    ResponseError(ResponseError),
    RootDirectoryAlreadyExists,
    RandomDataGenerationFailure,
    OperationForbiddenForClient,
    Unexpected(String),
    RoutingError(RoutingError),
    UnsupportedSaltSizeForPwHash,
    UnsuccessfulPwHash,
    OperationAborted,
}

Client Errors

Variants

StructuredDataHeaderSizeProhibitive

StructuredData has no space available to fit in any user data inside it.

UnsuccessfulEncodeDecode

Could not Serialise or Deserialise

AsymmetricDecipherFailure

Asymmetric Key Decryption Failed

SymmetricDecipherFailure

Symmetric Key Decryption Failed

ReceivedUnexpectedData

ReceivedUnexpectedData

VersionCacheMiss

No such data found in local version cache

RoutingMessageCacheMiss

No such data found in routing-filled cache

ResponseError

Network operation failed

RootDirectoryAlreadyExists

Cannot overwrite a root directory if it already exists

RandomDataGenerationFailure

Unable to obtain generator for random data

OperationForbiddenForClient

Forbidden operation requested for this Client

Unexpected

Unexpected - Probably a Logic error

RoutingError

Routing Error

UnsupportedSaltSizeForPwHash

Unable to pack into or operate with size of Salt

UnsuccessfulPwHash

Unable to complete computation for password hashing - usually because OS refused to allocate amount of requested memory

OperationAborted

Blocking operation was cancelled