Enum safe_core::ffi::errors::FfiError

pub enum FfiError {
    CoreError(Box<CoreError>),
    NfsError(Box<NfsError>),
    DnsError(Box<DnsError>),
    PathNotFound,
    InvalidPath,
    PermissionDenied,
    JsonParseError(ParserError),
    JsonDecodeError(DecoderError),
    SpecificParseError(String),
    JsonEncodeError(EncoderError),
    LocalConfigAccessFailed(String),
    Unexpected(String),
    UnsuccessfulEncodeDecode(SerialisationError),
}

Launcher Errors

Variants
CoreError(Box<CoreError>)

Error from safe_core. Boxed to hold a pointer instead of value so that this enum variant is not insanely bigger than others.

NfsError(Box<NfsError>)

Errors from safe_nfs

DnsError(Box<DnsError>)

Errors from safe_nfs

PathNotFound

Unable to find/traverse directory or file path

InvalidPath

Supplied path was invalid

PermissionDenied

Permission denied - e.g. permission to access SAFEDrive etc.

JsonParseError(ParserError)

Could not parse payload as a valid JSON

JsonDecodeError(DecoderError)

Could not decode valid JSON into expected Structures probably because a mandatory field was missing or a field was wrongly named etc.

SpecificParseError(String)

JSON non-conforming to the Launcher RFC and not covered by JsonDecodeError, e.g. things like invalid base64 formatting, unreasonable/unexpected indexing, ranges etc.

JsonEncodeError(EncoderError)

Error encoding into Json String

LocalConfigAccessFailed(String)

Unable to Read from or Write to a Local Config file.

Unexpected(String)

Unexpected - Probably a Logic error

UnsuccessfulEncodeDecode(SerialisationError)

Could not serialise or deserialise data