Module safe_core::ffi::low_level_api::immut_data

Low level manipulation of ImmutableData

Structs

SelfEncryptorReaderWrapper
SelfEncryptorWriterWrapper ties in the objects with dependent lifetimes and manages correct destruction sequence.

SelfEncryptorWriterWrapper
SelfEncryptorWriterWrapper ties in the objects with dependent lifetimes and manages correct destruction sequence.

Functions

immut_data_close_self_encryptor
Close Self Encryptor

immut_data_fetch_self_encryptor
Fetch Self Encryptor

immut_data_new_self_encryptor
Get a Self Encryptor

immut_data_read_from_self_encryptor
Read from Self Encryptor

immut_data_self_encryptor_reader_free
Free Self Encryptor Reader handle

immut_data_self_encryptor_writer_free
Free Self Encryptor Writer handle

immut_data_size
Get data size from Self Encryptor

immut_data_write_to_self_encryptor
Write to Self Encryptor