Module safe_core::ffi::string_list

StringList is a FFI-enabled wrapper over vector of strings with a corresponding FFI-enabled API.

Functions

into_ptr
Convert vec of strings to an owning raw pointer to StringList. Consumes the vector.

string_list_at
Get the string at the given index, or NULL if the index is out of range.

string_list_drop
Dispose of the string list.

string_list_len
Get number of elements in the string list.

Type Definitions

StringList
List of strings.