Function safe_core::structured_data_operations::unversioned::create

pub fn create(client: Arc<Mutex<Client>>, tag_type: u64, id: XorName, version: u64, data: Vec<u8>, owner_keys: Vec<PublicKey>, prev_owner_keys: Vec<PublicKey>, private_signing_key: &SecretKey, data_encryption_keys: Option<(&PublicKey, &SecretKey, &Nonce)>) -> Result<StructuredData, CoreError>

Create StructuredData in accordance with data-encoding rules abstracted from user. For StructuredData created with create, data must be obtained using the complementary function defined in this module to get_data()