Module safe_core::translated_events

Events filtered from set of Routing provided events, on which the Client Modules must specifically act

Enums

DataReceivedEvent
Reception of reqested Data

NetworkEvent
Netowork Events that Client Modules need to deal with

OperationFailureEvent
Failures in operations that Client Modules need to deal with