Crate safe_core

Safe-Core Library

Project github page

Reexports

pub use self_encryption_storage::SelfEncryptionStorage;

Modules

client
Self-Authentication and Gateway Module

errors
Safe-Core Errors

id
Public and Private Id types

self_encryption_storage
Implements the Self Encryption storage trait

structured_data_operations
Helper functions to handle StructuredData related operations

translated_events
Events filtered from set of Routing provided events, on which the Client Modules must specifically act

utility
Utility functions

Constants

CLIENT_STRUCTURED_DATA_TAG
All StructuredData tagging should positive-offset from this if the operation needs to go through this safe_core crate

MAIDSAFE_TAG
All Maidsafe tagging should positive-offset from this