Enum safe_core::ffi::errors::FfiError

pub enum FfiError {
    CoreError(Box<CoreError>),
    NfsError(Box<NfsError>),
    DnsError(Box<DnsError>),
    PathNotFound,
    InvalidPath,
    PermissionDenied,
    LocalConfigAccessFailed(String),
    Unexpected(String),
    UnsuccessfulEncodeDecode(SerialisationError),
    NulError(NulError),
    InvalidStructDataHandle,
    InvalidDataIdHandle,
    InvalidAppendableDataHandle,
    InvalidSelfEncryptorHandle,
    InvalidCipherOptHandle,
    InvalidEncryptKeyHandle,
    InvalidSignKeyHandle,
    OperationForbiddenForApp,
    InvalidStructuredDataTypeTag,
    InvalidVersionNumber,
    InvalidSelfEncryptorReadOffsets,
    InvalidIndex,
    UnsupportedOperation,
}

Launcher Errors

Variants
CoreError(Box<CoreError>)

Error from safe_core. Boxed to hold a pointer instead of value so that this enum variant is not insanely bigger than others.

NfsError(Box<NfsError>)

Errors from safe_nfs

DnsError(Box<DnsError>)

Errors from safe_nfs

PathNotFound

Unable to find/traverse directory or file path

InvalidPath

Supplied path was invalid

PermissionDenied

Permission denied - e.g. permission to access SAFEDrive etc.

LocalConfigAccessFailed(String)

Unable to Read from or Write to a Local Config file.

Unexpected(String)

Unexpected - Probably a Logic error

UnsuccessfulEncodeDecode(SerialisationError)

Could not serialise or deserialise data

NulError(NulError)

Could not convert String to nul-terminated string because it contains internal nuls.

InvalidStructDataHandle

Invalid StructuredData handle

InvalidDataIdHandle

Invalid DataIdentifier handle

InvalidAppendableDataHandle

Invalid Pub/PrivAppendableData handle

InvalidSelfEncryptorHandle

Invalid Self Encryptor handle

InvalidCipherOptHandle

Invalid CipherOpt handle

InvalidEncryptKeyHandle

Invalid encrypt (box_) key handle

InvalidSignKeyHandle

Invalid sign key handle

OperationForbiddenForApp

The requested operation is forbidded for the given app.

InvalidStructuredDataTypeTag

Invalid type tag for StructuredData

InvalidVersionNumber

Invalid version number requested for a versioned StructuredData

InvalidSelfEncryptorReadOffsets

Invalid offsets (from-position and lenght combination) provided for reading form Self Encryptor. Would have probably caused an overflow.

InvalidIndex

Invalid indexing

UnsupportedOperation

Unsupported Operation (e.g. mixing Pub/PrivAppendableData operations